+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/workshop/WorkshopsList.kt
@@ -4,15 +4,23 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
+import androidx.compose.material3.SnackbarHost
+import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavHostController
 import com.app.arcabyolimpo.presentation.navigation.Screen
+import com.app.arcabyolimpo.presentation.ui.components.atoms.alerts.Snackbarcustom
 import com.app.arcabyolimpo.presentation.ui.components.atoms.buttons.AddButton
 import com.app.arcabyolimpo.ui.theme.Background
+import kotlinx.coroutines.launch
 
 /**
  * Composable screen that displays the list of workshops.
@@ -33,9 +41,34 @@ fun WorkshopsListScreen(
     viewModel: WorkshopsListViewModel = hiltViewModel()
 )  {
     val uiState by viewModel.uiState.collectAsStateWithLifecycle()
+    val snackbarHostState = remember { SnackbarHostState() }
+    val scope = rememberCoroutineScope()
+
+    val savedStateHandle = navController.currentBackStackEntry?.savedStateHandle
+    val snackbarMessage = savedStateHandle?.get<String>("snackbarMessage")
+    val snackbarSuccess = savedStateHandle?.get<Boolean>("snackbarSuccess") ?: true
+
+    LaunchedEffect(snackbarMessage) {
+        snackbarMessage?.let {
+            scope.launch {
+                snackbarHostState.showSnackbar(it)
+            }
+            savedStateHandle.remove<String>("snackbarMessage")
+            savedStateHandle.remove<Boolean>("snackbarSuccess")
+        }
+    }
 
     Scaffold(
         containerColor = Background,
+        snackbarHost = {
+            SnackbarHost(snackbarHostState) { data ->
+                Snackbarcustom(
+                    title = data.visuals.message,
+                    modifier = Modifier.padding(16.dp),
+                    ifSucces = snackbarSuccess
+                )
+            }
+        },
         floatingActionButton = {
             AddButton(
                 onClick = {
@@ -43,7 +76,7 @@ fun WorkshopsListScreen(
                 }
             )
         }
-    ) { padding ->
+    ){ padding ->
         Box(
             modifier = Modifier
                 .fillMaxSize()